+++ src/main/java/kitchenpos/application/MenuService.java
@@ -9,10 +9,10 @@ import java.util.List;
 import kitchenpos.application.exception.CustomIllegalArgumentException;
 import kitchenpos.dao.JpaMenuGroupRepository;
 import kitchenpos.dao.JpaMenuRepository;
-import kitchenpos.dao.JpaProductRepository;
+import kitchenpos.product.domain.JpaProductRepository;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 import kitchenpos.ui.dto.request.MenuProductRequest;
 import kitchenpos.ui.dto.request.MenuRequest;
 import org.springframework.stereotype.Service;
