+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,11 +1,11 @@
-package kitchenpos.application;
+package kitchenpos.product.application;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.dao.JpaProductRepository;
-import kitchenpos.domain.Product;
-import kitchenpos.ui.dto.ProductResponse;
-import kitchenpos.ui.dto.request.ProductRequest;
+import kitchenpos.product.domain.JpaProductRepository;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.application.response.ProductResponse;
+import kitchenpos.product.ui.request.ProductRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
