+++ src/main/java/kitchenpos/product/application/response/ProductResponse.java
@@ -1,7 +1,7 @@
-package kitchenpos.ui.dto;
+package kitchenpos.product.application.response;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 
 public class ProductResponse {
