+++ src/main/java/kitchenpos/product/domain/JpaProductRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.dao;
+package kitchenpos.product.domain;
 
-import kitchenpos.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaProductRepository extends JpaRepository<Product, Long> {
