+++ src/test/java/kitchenpos/application/ServiceTest.java
@@ -13,7 +13,7 @@ import kitchenpos.dao.JpaMenuGroupRepository;
 import kitchenpos.dao.JpaMenuRepository;
 import kitchenpos.dao.JpaOrderRepository;
 import kitchenpos.dao.JpaOrderTableRepository;
-import kitchenpos.dao.JpaProductRepository;
+import kitchenpos.product.domain.JpaProductRepository;
 import kitchenpos.dao.JpaTableGroupRepository;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuProduct;
@@ -21,7 +21,7 @@ import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 import kitchenpos.domain.TableGroup;
 import kitchenpos.ui.dto.request.OrderLineItemRequest;
 import kitchenpos.ui.dto.request.OrderRequest;
