+++ src/test/java/kitchenpos/ui/ProductRestControllerTest.java
@@ -13,8 +13,8 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
-import kitchenpos.ui.dto.ProductResponse;
-import kitchenpos.ui.dto.request.ProductRequest;
+import kitchenpos.product.application.response.ProductResponse;
+import kitchenpos.product.ui.request.ProductRequest;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MvcResult;
